Rydder litt og legger til rette for å jobbe i parallell
