+++ src/app/personside/infotabs/meldinger/traadvisning/TraadVisning.tsx
@@ -5,7 +5,7 @@ import { datoSynkende } from '../../../../../utils/dateUtils';
 import EnkeltMelding from './Enkeltmelding';
 import AlertStripeInfo from 'nav-frontend-alertstriper/lib/info-alertstripe';
 import theme from '../../../../../styles/personOversiktTheme';
-import ToppMenyContainer from '../toppmeny/ToppMenyContainer';
+import VerktoylinjeContainer from './verktoylinje/VerktoylinjeContainer';
 
 interface Props {
     valgtTraad?: Traad;
@@ -36,7 +36,7 @@ class TraadVisning extends React.PureComponent<Props> {
 
         return (
             <VisningStyle aria-label={'Meldinger for valgt tråd'}>
-                <ToppMenyContainer />
+                <VerktoylinjeContainer />
                 <AlleMeldinger traad={this.props.valgtTraad} />
             </VisningStyle>
         );
