+++ src/app/personside/infotabs/meldinger/traadvisning/verktoylinje/journalforing/JournalforingPanel.tsx
